Make change button functional
Insert prompt that allows for user to input desired amount of boxes (max 100),
along with safety conditions preventing them from inserting negative, null
or non-numerical values.

Also insert removeChild code to remove default grid and reused the
same code to recreate a new grid, make it sketchable and have the
the width and height dimensions of the boxes scale accordingly to the
size of the container and the number of boxes.